requirements.txt
fastapi==0.108.0
sqlmodel==0.0.16
psycopg2-binary==2.9.9
python-dotenv==1.0.0
uvicorn==0.25.0
pytest==8.0.0
httpx==0.26.0
alembic==1.13.1
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9